// server/src/types/radar.ts
export interface ObjectData {
  deviceId: string;
  frameType: '0x01';
  timestamp: string;
  numEntries: number;
  entries: VehicleEntry[];
  packetSize: number;
}

export interface VehicleEntry {
  targetId: string;
  laneNo: number;
  targetType: number;
  color: number;
  plateNumber: string;
  xCoordM: number;
  yCoordM: number;
  speedKmh: number;
  azimuthDeg: number;
  longitude: number;
  latitude: number;
  imageX: number;
  imageY: number;
  vehicleLength: number;
  vehicleWidth: number;
  vehicleHeight: number;
  parkingStatus: boolean;
  xSpeed: number;
  ySpeed: number;
  acceleration: number;
}